@@ -2,11 +2,9 @@ package codechicken.lib.block.component.data;
 
 import codechicken.lib.block.ModularBlock;
 import codechicken.lib.datagen.NoValidationBLockLootSubProvider;
+import net.minecraft.core.HolderLookup;
 import net.minecraft.data.DataProvider;
-import net.minecraft.data.loot.BlockLootSubProvider;
 import net.minecraft.data.loot.LootTableProvider;
-import net.minecraft.world.flag.FeatureFlagSet;
-import net.minecraft.world.flag.FeatureFlags;
 import net.minecraft.world.level.ItemLike;
 import net.minecraft.world.level.storage.loot.LootPool;
 import net.minecraft.world.level.storage.loot.LootTable;
@@ -16,7 +14,6 @@ import net.minecraft.world.level.storage.loot.providers.number.ConstantValue;
 import org.jetbrains.annotations.ApiStatus;
 
 import java.util.ArrayList;
-import java.util.Set;
 import java.util.function.Consumer;
 
 /**
@@ -42,7 +39,7 @@ public class LootTableComponent extends DataGenComponent {
             }
             provider.subProviders.add(
                     new LootTableProvider.SubProviderEntry(
-                            () -> new NoValidationBLockLootSubProvider() {
+                            registries -> new NoValidationBLockLootSubProvider(registries) {
                                 @Override
                                 protected void generate() {
                                     add(getBlock(), table);
