@@ -3,10 +3,14 @@ package codechicken.lib.colour;
 import com.google.common.collect.HashBasedTable;
 import com.google.common.collect.ImmutableTable;
 import com.google.common.collect.Table;
+import io.netty.buffer.ByteBuf;
 import net.covers1624.quack.collection.FastStream;
 import net.minecraft.client.resources.language.I18n;
+import net.minecraft.network.codec.ByteBufCodecs;
+import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.TagKey;
+import net.minecraft.util.ByIdMap;
 import net.minecraft.util.StringRepresentable;
 import net.minecraft.world.item.ItemStack;
 import org.jetbrains.annotations.Nullable;
@@ -14,6 +18,7 @@ import org.jetbrains.annotations.Nullable;
 import java.util.Map;
 import java.util.Objects;
 import java.util.function.Function;
+import java.util.function.IntFunction;
 
 /**
  * Created by covers1624 on 16/09/2016.
@@ -21,23 +26,27 @@ import java.util.function.Function;
 public enum EnumColour implements StringRepresentable {
 
     //@formatter:off
-    WHITE     ("white",      "forge:dyes/white",      "forge:wool/white",      "item.minecraft.firework_star.white",       0xFFFFFF),
-    ORANGE    ("orange",     "forge:dyes/orange",     "forge:wool/orange",     "item.minecraft.firework_star.orange",      0xC06300),
-    MAGENTA   ("magenta",    "forge:dyes/magenta",    "forge:wool/magenta",    "item.minecraft.firework_star.magenta",     0xB51AB5),
-    LIGHT_BLUE("light_blue", "forge:dyes/light_blue", "forge:wool/light_blue", "item.minecraft.firework_star.light_blue",  0x6F84F1),
-    YELLOW    ("yellow",     "forge:dyes/yellow",     "forge:wool/yellow",     "item.minecraft.firework_star.yellow",      0xBFBF00),
-    LIME      ("lime",       "forge:dyes/lime",       "forge:wool/lime",       "item.minecraft.firework_star.lime",        0x6BF100),
-    PINK      ("pink",       "forge:dyes/pink",       "forge:wool/pink",       "item.minecraft.firework_star.pink",        0xF14675),
-    GRAY      ("gray",       "forge:dyes/gray",       "forge:wool/gray",       "item.minecraft.firework_star.gray",        0x535353),
-    LIGHT_GRAY("light_gray", "forge:dyes/light_gray", "forge:wool/light_gray", "item.minecraft.firework_star.light_gray",  0x939393),
-    CYAN      ("cyan",       "forge:dyes/cyan",       "forge:wool/cyan",       "item.minecraft.firework_star.cyan",        0x008787),
-    PURPLE    ("purple",     "forge:dyes/purple",     "forge:wool/purple",     "item.minecraft.firework_star.purple",      0x5E00C0),
-    BLUE      ("blue",       "forge:dyes/blue",       "forge:wool/blue",       "item.minecraft.firework_star.blue",        0x1313C0),
-    BROWN     ("brown",      "forge:dyes/brown",      "forge:wool/brown",      "item.minecraft.firework_star.brown",       0x4F2700),
-    GREEN     ("green",      "forge:dyes/green",      "forge:wool/green",      "item.minecraft.firework_star.green",       0x088700),
-    RED       ("red",        "forge:dyes/red",        "forge:wool/red",        "item.minecraft.firework_star.red",         0xA20F06),
-    BLACK     ("black",      "forge:dyes/black",      "forge:wool/black",      "item.minecraft.firework_star.black",       0x1F1F1F);
+    WHITE     ("white",      "c:dyes/white",      "c:wools/white",      "item.minecraft.firework_star.white",       0xFFFFFF),
+    ORANGE    ("orange",     "c:dyes/orange",     "c:wools/orange",     "item.minecraft.firework_star.orange",      0xC06300),
+    MAGENTA   ("magenta",    "c:dyes/magenta",    "c:wools/magenta",    "item.minecraft.firework_star.magenta",     0xB51AB5),
+    LIGHT_BLUE("light_blue", "c:dyes/light_blue", "c:wools/light_blue", "item.minecraft.firework_star.light_blue",  0x6F84F1),
+    YELLOW    ("yellow",     "c:dyes/yellow",     "c:wools/yellow",     "item.minecraft.firework_star.yellow",      0xBFBF00),
+    LIME      ("lime",       "c:dyes/lime",       "c:wools/lime",       "item.minecraft.firework_star.lime",        0x6BF100),
+    PINK      ("pink",       "c:dyes/pink",       "c:wools/pink",       "item.minecraft.firework_star.pink",        0xF14675),
+    GRAY      ("gray",       "c:dyes/gray",       "c:wools/gray",       "item.minecraft.firework_star.gray",        0x535353),
+    LIGHT_GRAY("light_gray", "c:dyes/light_gray", "c:wools/light_gray", "item.minecraft.firework_star.light_gray",  0x939393),
+    CYAN      ("cyan",       "c:dyes/cyan",       "c:wools/cyan",       "item.minecraft.firework_star.cyan",        0x008787),
+    PURPLE    ("purple",     "c:dyes/purple",     "c:wools/purple",     "item.minecraft.firework_star.purple",      0x5E00C0),
+    BLUE      ("blue",       "c:dyes/blue",       "c:wools/blue",       "item.minecraft.firework_star.blue",        0x1313C0),
+    BROWN     ("brown",      "c:dyes/brown",      "c:wools/brown",      "item.minecraft.firework_star.brown",       0x4F2700),
+    GREEN     ("green",      "c:dyes/green",      "c:wools/green",      "item.minecraft.firework_star.green",       0x088700),
+    RED       ("red",        "c:dyes/red",        "c:wools/red",        "item.minecraft.firework_star.red",         0xA20F06),
+    BLACK     ("black",      "c:dyes/black",      "c:wools/black",      "item.minecraft.firework_star.black",       0x1F1F1F);
     //@formatter:on
+
+    public static final StringRepresentable.EnumCodec<EnumColour> CODEC = StringRepresentable.fromEnum(EnumColour::values);
+    public static final IntFunction<EnumColour> BY_ID = ByIdMap.continuous(EnumColour::ordinal, values(), ByIdMap.OutOfBoundsStrategy.WRAP);
+    public static final StreamCodec<ByteBuf, EnumColour> STREAM_CODEC = ByteBufCodecs.idMapper(BY_ID, EnumColour::ordinal);
 
     private final String name;
     private final ResourceLocation dyeTagName;
@@ -87,8 +96,8 @@ public enum EnumColour implements StringRepresentable {
 
     EnumColour(String name, String dyeTagName, String woolTagName, String unlocalizedName, int rgb) {
         this.name = name;
-        this.dyeTagName = new ResourceLocation(dyeTagName);
-        this.woolTagName = new ResourceLocation(woolTagName);
+        this.dyeTagName = ResourceLocation.parse(dyeTagName);
+        this.woolTagName = ResourceLocation.parse(woolTagName);
         this.unlocalizedName = unlocalizedName;
         this.rgb = rgb;
     }
