@@ -2,10 +2,16 @@ package codechicken.lib.data;
 
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
+import net.minecraft.core.RegistryAccess;
 import net.minecraft.nbt.ByteArrayTag;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.Tag;
 import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.network.RegistryFriendlyByteBuf;
+import net.minecraft.network.codec.StreamDecoder;
+import net.minecraft.network.codec.StreamEncoder;
+import net.neoforged.neoforge.network.connection.ConnectionType;
+import org.jetbrains.annotations.Nullable;
 
 /**
  * An {@link MCDataInput} and {@link MCDataOutput} implementation,
@@ -16,13 +22,26 @@ import net.minecraft.network.FriendlyByteBuf;
 public class MCDataByteBuf implements MCDataInput, MCDataOutput {
 
     protected final ByteBuf buf;
+    protected final @Nullable RegistryAccess registryAccess;
 
     public MCDataByteBuf() {
-        this(Unpooled.buffer());
+        this(Unpooled.buffer(), null);
+    }
+
+    public MCDataByteBuf(@Nullable RegistryAccess registryAccess) {
+        this(Unpooled.buffer(), registryAccess);
     }
 
     public MCDataByteBuf(ByteBuf buf) {
+        this(buf, null);
+    }
+
+    public MCDataByteBuf(ByteBuf buf, @Nullable RegistryAccess registryAccess) {
         this.buf = buf;
+        if (registryAccess == null && buf instanceof RegistryFriendlyByteBuf rBuf) {
+            registryAccess = rBuf.registryAccess();
+        }
+        this.registryAccess = registryAccess;
     }
 
     /**
@@ -34,15 +53,23 @@ public class MCDataByteBuf implements MCDataInput, MCDataOutput {
         return buf instanceof FriendlyByteBuf ? (FriendlyByteBuf) buf : new FriendlyByteBuf(buf);
     }
 
+    public RegistryFriendlyByteBuf toRegistryFriendlyByteBuf() {
+        if (buf instanceof RegistryFriendlyByteBuf rBuf) return rBuf;
+        if (registryAccess == null) {
+            throw new RuntimeException("RegistryAccess required for this operation.");
+        }
+        return new RegistryFriendlyByteBuf(buf, registryAccess, ConnectionType.NEOFORGE);
+    }
+
     public Tag toTag() {
         return new ByteArrayTag(buf.array());
     }
 
-    public static MCDataByteBuf fromTag(Tag tag) {
+    public static MCDataByteBuf fromTag(Tag tag, RegistryAccess registries) {
         if (!(tag instanceof ByteArrayTag)) {
             throw new IllegalArgumentException("Expected ByteArrayNBT, got: " + tag.getClass().getSimpleName());
         }
-        return new MCDataByteBuf(Unpooled.copiedBuffer(((ByteArrayTag) tag).getAsByteArray()));
+        return new MCDataByteBuf(Unpooled.copiedBuffer(((ByteArrayTag) tag).getAsByteArray()), registries);
     }
 
     public CompoundTag writeToNBT(CompoundTag tag, String key) {
@@ -50,8 +77,30 @@ public class MCDataByteBuf implements MCDataInput, MCDataOutput {
         return tag;
     }
 
-    public static MCDataByteBuf readFromNBT(CompoundTag tag, String key) {
-        return fromTag(tag.get(key));
+    public static MCDataByteBuf readFromNBT(CompoundTag tag, String key, RegistryAccess registries) {
+        return fromTag(tag.get(key), registries);
+    }
+
+    @Override
+    public <T> MCDataOutput writeWithCodec(StreamEncoder<? super FriendlyByteBuf, T> codec, T thing) {
+        codec.encode(toFriendlyByteBuf(), thing);
+        return this;
+    }
+
+    @Override
+    public <T> MCDataOutput writeWithRegistryCodec(StreamEncoder<? super RegistryFriendlyByteBuf, T> codec, T thing) {
+        codec.encode(toRegistryFriendlyByteBuf(), thing);
+        return this;
+    }
+
+    @Override
+    public <T> T readWithCodec(StreamDecoder<? super FriendlyByteBuf, T> codec) {
+        return codec.decode(toFriendlyByteBuf());
+    }
+
+    @Override
+    public <T> T readWithRegistryCodec(StreamDecoder<? super RegistryFriendlyByteBuf, T> codec) {
+        return codec.decode(toRegistryFriendlyByteBuf());
     }
 
     //@formatter:off
