@@ -1,6 +1,9 @@
 package codechicken.lib.datagen;
 
+import net.minecraft.core.HolderLookup;
+import net.minecraft.core.RegistryAccess;
 import net.minecraft.data.loot.BlockLootSubProvider;
+import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.flag.FeatureFlagSet;
 import net.minecraft.world.item.Item;
@@ -16,27 +19,27 @@ import java.util.function.BiConsumer;
  */
 public abstract class NoValidationBLockLootSubProvider extends BlockLootSubProvider {
 
-    private final Map<ResourceLocation, LootTable.Builder> map;
+    private final Map<ResourceKey<LootTable>, LootTable.Builder> map;
 
-    protected NoValidationBLockLootSubProvider() {
-        this(Set.of());
+    protected NoValidationBLockLootSubProvider(HolderLookup.Provider registries) {
+        this(Set.of(), registries);
     }
 
-    protected NoValidationBLockLootSubProvider(Set<Item> explosionResistant) {
-        this(explosionResistant, FeatureFlagSet.of());
+    protected NoValidationBLockLootSubProvider(Set<Item> explosionResistant, HolderLookup.Provider registries) {
+        this(explosionResistant, FeatureFlagSet.of(), registries);
     }
 
-    protected NoValidationBLockLootSubProvider(Set<Item> explosionResistant, FeatureFlagSet flags) {
-        this(explosionResistant, flags, new HashMap<>());
+    protected NoValidationBLockLootSubProvider(Set<Item> explosionResistant, FeatureFlagSet flags, HolderLookup.Provider registries) {
+        this(explosionResistant, flags, new HashMap<>(), registries);
     }
 
-    protected NoValidationBLockLootSubProvider(Set<Item> explosionResistant, FeatureFlagSet flags, Map<ResourceLocation, LootTable.Builder> map) {
-        super(explosionResistant, flags, map);
+    protected NoValidationBLockLootSubProvider(Set<Item> explosionResistant, FeatureFlagSet flags, Map<ResourceKey<LootTable>, LootTable.Builder> map, HolderLookup.Provider registries) {
+        super(explosionResistant, flags, map, registries);
         this.map = map;
     }
 
     @Override
-    public void generate(BiConsumer<ResourceLocation, LootTable.Builder> cons) {
+    public void generate(BiConsumer<ResourceKey<LootTable>, LootTable.Builder> cons) {
         generate();
         map.forEach(cons);
     }
