@@ -26,11 +26,11 @@ import java.util.Optional;
 public class CursorHelper {
     public static final Logger LOGGER = LogManager.getLogger();
 
-    public static final ResourceLocation DRAG = new ResourceLocation(CodeChickenLib.MOD_ID, "textures/gui/cursors/drag.png");
-    public static final ResourceLocation RESIZE_H = new ResourceLocation(CodeChickenLib.MOD_ID, "textures/gui/cursors/resize_h.png");
-    public static final ResourceLocation RESIZE_V = new ResourceLocation(CodeChickenLib.MOD_ID, "textures/gui/cursors/resize_v.png");
-    public static final ResourceLocation RESIZE_TRBL = new ResourceLocation(CodeChickenLib.MOD_ID, "textures/gui/cursors/resize_diag_trbl.png");
-    public static final ResourceLocation RESIZE_TLBR = new ResourceLocation(CodeChickenLib.MOD_ID, "textures/gui/cursors/resize_diag_tlbr.png");
+    public static final ResourceLocation DRAG = ResourceLocation.fromNamespaceAndPath(CodeChickenLib.MOD_ID, "textures/gui/cursors/drag.png");
+    public static final ResourceLocation RESIZE_H = ResourceLocation.fromNamespaceAndPath(CodeChickenLib.MOD_ID, "textures/gui/cursors/resize_h.png");
+    public static final ResourceLocation RESIZE_V = ResourceLocation.fromNamespaceAndPath(CodeChickenLib.MOD_ID, "textures/gui/cursors/resize_v.png");
+    public static final ResourceLocation RESIZE_TRBL = ResourceLocation.fromNamespaceAndPath(CodeChickenLib.MOD_ID, "textures/gui/cursors/resize_diag_trbl.png");
+    public static final ResourceLocation RESIZE_TLBR = ResourceLocation.fromNamespaceAndPath(CodeChickenLib.MOD_ID, "textures/gui/cursors/resize_diag_tlbr.png");
 
     private static final Map<ResourceLocation, Long> cursors = new HashMap<>();
     private static ResourceLocation active = null;
