@@ -1,8 +1,8 @@
 package codechicken.lib.internal;
 
 import net.covers1624.quack.util.CrashLock;
+import net.neoforged.neoforge.client.event.ClientTickEvent;
 import net.neoforged.neoforge.common.NeoForge;
-import net.neoforged.neoforge.event.TickEvent;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -23,15 +23,12 @@ public class ExceptionMessageEventHandler {
         NeoForge.EVENT_BUS.addListener(ExceptionMessageEventHandler::clientTick);
     }
 
-    private static void clientTick(TickEvent.ClientTickEvent event) {
-        if (event.phase == TickEvent.Phase.END) {
-            //Clear the known exceptions every 5 seconds.
-            long time = System.nanoTime();
-            if (TimeUnit.NANOSECONDS.toSeconds(time - lastExceptionClear) > 5) {
-                lastExceptionClear = time;
-                exceptionMessageCache.clear();
-            }
+    private static void clientTick(ClientTickEvent.Post event) {
+        //Clear the known exceptions every 5 seconds.
+        long time = System.nanoTime();
+        if (TimeUnit.NANOSECONDS.toSeconds(time - lastExceptionClear) > 5) {
+            lastExceptionClear = time;
+            exceptionMessageCache.clear();
         }
     }
-
 }
