@@ -5,14 +5,19 @@ import codechicken.lib.packet.ICustomPacketHandler.IClientPacketHandler;
 import codechicken.lib.packet.ICustomPacketHandler.IServerConfigurationPacketHandler;
 import codechicken.lib.packet.ICustomPacketHandler.IServerPacketHandler;
 import net.minecraft.client.Minecraft;
+import net.minecraft.core.RegistryAccess;
 import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.network.RegistryFriendlyByteBuf;
+import net.minecraft.network.codec.StreamCodec;
+import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.loading.FMLEnvironment;
-import net.neoforged.neoforge.network.event.RegisterPayloadHandlerEvent;
-import net.neoforged.neoforge.network.handling.ISynchronizedWorkHandler;
-import net.neoforged.neoforge.network.registration.IPayloadRegistrar;
+import net.neoforged.neoforge.network.event.RegisterPayloadHandlersEvent;
+import net.neoforged.neoforge.network.handling.IPayloadContext;
+import net.neoforged.neoforge.network.registration.HandlerThread;
+import net.neoforged.neoforge.network.registration.PayloadRegistrar;
 import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -79,63 +84,70 @@ public class PacketCustomChannel {
         modBus.addListener(this::onRegisterPayloadHandlerEvent);
     }
 
-    private void onRegisterPayloadHandlerEvent(RegisterPayloadHandlerEvent event) {
-        IPayloadRegistrar registrar = event.registrar(channel.getNamespace());
+    private void onRegisterPayloadHandlerEvent(RegisterPayloadHandlersEvent event) {
+        PayloadRegistrar registrar = event.registrar(channel.getNamespace());
 
         if (optional) registrar = registrar.optional();
         if (version != null) registrar = registrar.versioned(version);
+        registrar = registrar.executesOn(HandlerThread.NETWORK);
 
-        registrar.play(
-                channel,
-                this::read,
-                handlers -> {
-                    handlers.client((payload, context) -> {
-                        if (client != null) {
-                            enqueue(context.workHandler(), payload, () -> {
-                                client.handlePacket(new PacketCustom(payload), Minecraft.getInstance());
-                            });
-                        }
-                    });
-                    handlers.server((payload, context) -> {
-                        if (server != null) {
-                            enqueue(context.workHandler(), payload, () -> {
-                                server.handlePacket(new PacketCustom(payload), (ServerPlayer) context.player().orElseThrow());
-                            });
-                        }
-                    });
+        CustomPacketPayload.Type<PacketCustom.Pkt> type = new CustomPacketPayload.Type<>(channel);
+        StreamCodec<FriendlyByteBuf, PacketCustom.Pkt> codec = StreamCodec.of(
+                (buf, p) -> {
+                    p.data().markReaderIndex();
+                    buf.writeBytes(p.data());
+                    p.data().resetReaderIndex();
+                },
+                buf -> {
+                    RegistryAccess access = buf instanceof RegistryFriendlyByteBuf b ? b.registryAccess() : null;
+                    return new PacketCustom.Pkt(type, access, buf.readBytes(buf.readableBytes()));
                 }
         );
-        registrar.configuration(
-                channel,
-                this::read,
-                handlers -> {
-                    handlers.client((payload, context) -> {
-                        if (clientConfiguration != null) {
-                            enqueue(context.workHandler(), payload, () -> {
-                                clientConfiguration.handlePacket(new PacketCustom(payload), Minecraft.getInstance());
-                            });
+        registrar.playBidirectional(
+                type,
+                codec,
+                (payload, context) -> {
+                    switch (context.flow()) {
+                        case CLIENTBOUND -> {
+                            if (client != null) {
+                                enqueue(context, payload, () -> {
+                                    client.handlePacket(new PacketCustom(payload), Minecraft.getInstance());
+                                });
+                            }
                         }
-                    });
+                        case SERVERBOUND -> {
+                            if (server != null) {
+                                enqueue(context, payload, () -> {
+                                    server.handlePacket(new PacketCustom(payload), (ServerPlayer) context.player());
+                                });
+                            }
+                        }
+                    }
+
+                }
+        );
+        registrar.configurationToClient(
+                type,
+                codec,
+                (payload, context) -> {
+                    if (clientConfiguration != null) {
+                        enqueue(context, payload, () -> {
+                            clientConfiguration.handlePacket(new PacketCustom(payload), Minecraft.getInstance());
+                        });
+                    }
                 }
         );
     }
 
-    private PacketCustom.Pkt read(FriendlyByteBuf buf) {
-        return new PacketCustom.Pkt(
-                channel,
-                buf.readBytes(buf.readableBytes())
-        );
-    }
-
-    private static void enqueue(ISynchronizedWorkHandler workHandler, PacketCustom.Pkt payload, Runnable runnable) {
-        workHandler.submitAsync(() -> {
+    private static void enqueue(IPayloadContext context, PacketCustom.Pkt payload, Runnable runnable) {
+        context.enqueueWork(() -> {
             try {
                 runnable.run();
             } finally {
                 payload.data().release();
             }
         }).exceptionally(ex -> {
-            LOGGER.error("Error processing packet on channel {}", payload.id(), ex);
+            LOGGER.error("Error processing packet on channel {}", payload.type().id(), ex);
             return null;
         });
     }
