@@ -298,7 +298,7 @@ public class OBJParser {
 
     private static ResourceLocation maybeRelative(ResourceLocation other, String resource) {
         if (resource.contains(":")) {
-            return new ResourceLocation(resource);
+            return ResourceLocation.parse(resource);
         }
         String path = other.getPath();
         int lastSlash = path.lastIndexOf("/");
@@ -307,6 +307,6 @@ public class OBJParser {
         } else {
             path = "";
         }
-        return new ResourceLocation(other.getNamespace(), path + "/" + resource);
+        return ResourceLocation.fromNamespaceAndPath(other.getNamespace(), path + "/" + resource);
     }
 }
