@@ -169,9 +169,9 @@ public class GlslProcessor {
                     if (!includeFolder) {
                         match = matcher.group(2);
                     }
-                    ResourceLocation loc = new ResourceLocation(match);
+                    ResourceLocation loc = ResourceLocation.parse(match);
                     if (includeFolder) {
-                        loc = new ResourceLocation(loc.getNamespace(), FilenameUtils.normalize("shaders/include/" + loc.getPath(), true));
+                        loc = ResourceLocation.fromNamespaceAndPath(loc.getNamespace(), FilenameUtils.normalize("shaders/include/" + loc.getPath(), true));
                     }
                     imports.add(loc);
                 }
