@@ -14,51 +14,51 @@ public class CCLTags {
 
     public static class Blocks {
 
-        public static TagKey<Block> WOOL = forge("wool");
-        public static TagKey<Block> WOOL_WHITE = forge("wool/white");
-        public static TagKey<Block> WOOL_ORANGE = forge("wool/orange");
-        public static TagKey<Block> WOOL_MAGENTA = forge("wool/magenta");
-        public static TagKey<Block> WOOL_LIGHT_BLUE = forge("wool/light_blue");
-        public static TagKey<Block> WOOL_YELLOW = forge("wool/yellow");
-        public static TagKey<Block> WOOL_LIME = forge("wool/lime");
-        public static TagKey<Block> WOOL_PINK = forge("wool/pink");
-        public static TagKey<Block> WOOL_GRAY = forge("wool/gray");
-        public static TagKey<Block> WOOL_LIGHT_GRAY = forge("wool/light_gray");
-        public static TagKey<Block> WOOL_CYAN = forge("wool/cyan");
-        public static TagKey<Block> WOOL_PURPLE = forge("wool/purple");
-        public static TagKey<Block> WOOL_BLUE = forge("wool/blue");
-        public static TagKey<Block> WOOL_BROWN = forge("wool/brown");
-        public static TagKey<Block> WOOL_GREEN = forge("wool/green");
-        public static TagKey<Block> WOOL_RED = forge("wool/red");
-        public static TagKey<Block> WOOL_BLACK = forge("wool/black");
+        public static TagKey<Block> WOOLS = common("wools");
+        public static TagKey<Block> WOOLS_WHITE = common("wools/white");
+        public static TagKey<Block> WOOLS_ORANGE = common("wools/orange");
+        public static TagKey<Block> WOOLS_MAGENTA = common("wools/magenta");
+        public static TagKey<Block> WOOLS_LIGHT_BLUE = common("wools/light_blue");
+        public static TagKey<Block> WOOLS_YELLOW = common("wools/yellow");
+        public static TagKey<Block> WOOLS_LIME = common("wools/lime");
+        public static TagKey<Block> WOOLS_PINK = common("wools/pink");
+        public static TagKey<Block> WOOLS_GRAY = common("wools/gray");
+        public static TagKey<Block> WOOLS_LIGHT_GRAY = common("wools/light_gray");
+        public static TagKey<Block> WOOLS_CYAN = common("wools/cyan");
+        public static TagKey<Block> WOOLS_PURPLE = common("wools/purple");
+        public static TagKey<Block> WOOLS_BLUE = common("wools/blue");
+        public static TagKey<Block> WOOLS_BROWN = common("wools/brown");
+        public static TagKey<Block> WOOLS_GREEN = common("wools/green");
+        public static TagKey<Block> WOOLS_RED = common("wools/red");
+        public static TagKey<Block> WOOLS_BLACK = common("wools/black");
 
-        private static TagKey<Block> forge(String path) {
-            return BlockTags.create(new ResourceLocation("forge", path));
+        private static TagKey<Block> common(String path) {
+            return BlockTags.create(ResourceLocation.fromNamespaceAndPath("c", path));
         }
     }
 
     public static class Items {
 
-        public static TagKey<Item> WOOL = forge("wool");
-        public static TagKey<Item> WOOL_WHITE = forge("wool/white");
-        public static TagKey<Item> WOOL_ORANGE = forge("wool/orange");
-        public static TagKey<Item> WOOL_MAGENTA = forge("wool/magenta");
-        public static TagKey<Item> WOOL_LIGHT_BLUE = forge("wool/light_blue");
-        public static TagKey<Item> WOOL_YELLOW = forge("wool/yellow");
-        public static TagKey<Item> WOOL_LIME = forge("wool/lime");
-        public static TagKey<Item> WOOL_PINK = forge("wool/pink");
-        public static TagKey<Item> WOOL_GRAY = forge("wool/gray");
-        public static TagKey<Item> WOOL_LIGHT_GRAY = forge("wool/light_gray");
-        public static TagKey<Item> WOOL_CYAN = forge("wool/cyan");
-        public static TagKey<Item> WOOL_PURPLE = forge("wool/purple");
-        public static TagKey<Item> WOOL_BLUE = forge("wool/blue");
-        public static TagKey<Item> WOOL_BROWN = forge("wool/brown");
-        public static TagKey<Item> WOOL_GREEN = forge("wool/green");
-        public static TagKey<Item> WOOL_RED = forge("wool/red");
-        public static TagKey<Item> WOOL_BLACK = forge("wool/black");
+        public static TagKey<Item> WOOLS = common("wools");
+        public static TagKey<Item> WOOLS_WHITE = common("wools/white");
+        public static TagKey<Item> WOOLS_ORANGE = common("wools/orange");
+        public static TagKey<Item> WOOLS_MAGENTA = common("wools/magenta");
+        public static TagKey<Item> WOOLS_LIGHT_BLUE = common("wools/light_blue");
+        public static TagKey<Item> WOOLS_YELLOW = common("wools/yellow");
+        public static TagKey<Item> WOOLS_LIME = common("wools/lime");
+        public static TagKey<Item> WOOLS_PINK = common("wools/pink");
+        public static TagKey<Item> WOOLS_GRAY = common("wools/gray");
+        public static TagKey<Item> WOOLS_LIGHT_GRAY = common("wools/light_gray");
+        public static TagKey<Item> WOOLS_CYAN = common("wools/cyan");
+        public static TagKey<Item> WOOLS_PURPLE = common("wools/purple");
+        public static TagKey<Item> WOOLS_BLUE = common("wools/blue");
+        public static TagKey<Item> WOOLS_BROWN = common("wools/brown");
+        public static TagKey<Item> WOOLS_GREEN = common("wools/green");
+        public static TagKey<Item> WOOLS_RED = common("wools/red");
+        public static TagKey<Item> WOOLS_BLACK = common("wools/black");
 
-        private static TagKey<Item> forge(String path) {
-            return ItemTags.create(new ResourceLocation("forge", path));
+        private static TagKey<Item> common(String path) {
+            return ItemTags.create(ResourceLocation.fromNamespaceAndPath("c", path));
         }
     }
 }
